Refactor chunk world positioning and clean up Transvoxel code
Transvoxels are still not working correctly in the current implementation.
At the moment, transitions only work for LOD4 -> LOD3. I was unable to
identify the root cause yet and suspect the issue may be elsewhere.

This commit focuses on cleanup and correctness so the code is easier to
reason about later when debugging with a fresh head.

IMPORTANT CHANGES:
- Chunk origin world positioning has been reworked.
  ChunkOriginToWorld now requires the chunk LOD:
    ChunkOriginToWorld(int3 origin, uint lodIndex)

- Cell world position math was updated:
  From:
    CellWorldPos = ChunkOriginToWorld(origin) + CellCoord * CellStep
    (origin implicitly assumed LOD0 spacing)
  To:
    CellWorldPos = ChunkOriginToWorld(chunk) + CellCoord * CellStep
    (origin now scaled by the chunk's actual LOD)

- ChunkKey.Origin is now treated as a logical chunk index instead of a
  world-space position baked with LOD scaling. Origins can now have
  distinct values per LOD.

NOTE:
LodOctTree has NOT yet been updated to account for this change in origin
handling. That will need to be addressed separately once Transvoxels
are fully understood and working.
