--- Terrain/Engine/Shaders/Compute/MarchingCubes.compute
@@ -1,209 +1,0 @@
-#pragma only_renderers d3d11
-#pragma target 5.0
-
-#pragma kernel RunMarchingCubesPrePass
-#pragma kernel RunMarchingCubes
-
-#include "../ChunkFunctions.hlsl"
-#include "../Includes/TriangleTable.hlsl"
-
-/* Inputs */
-uint Offset;
-StructuredBuffer<ChunkWorkDescriptor> ChunkInputs;
-StructuredBuffer<float> DensityMap;
-
-/* Outputs */
-RWStructuredBuffer<uint> TriangleCount;
-RWStructuredBuffer<uint> TriangleCursor;
-RWStructuredBuffer<TriangleData> TriangleSourceBuffer;
-
-float3 GetNormalFast(int3 p, int ChunkKeyIndex, int3 sampleSize)
-{
-    // Precompute strides in the flattened density buffer
-    const int strideX = 1;
-    const int strideY = sampleSize.x;
-    const int strideZ = sampleSize.x * sampleSize.y;
-
-    // Convert center sample position to flat index (handles borders)
-    int centerIndex = GetDensitySampleIndexPadded(p, ChunkKeyIndex, sampleSize);
-
-    // Central differences
-    float dx = DensityMap[centerIndex + strideX] - DensityMap[centerIndex - strideX];
-    float dy = DensityMap[centerIndex + strideY] - DensityMap[centerIndex - strideY];
-    float dz = DensityMap[centerIndex + strideZ] - DensityMap[centerIndex - strideZ];
-
-    float3 n = -float3(dx, dy, dz);
-   
-    return n;
-}
-
-void CountTriangles(ChunkCellContext key)
-{
-    uint CellCoord = 0;
-    int3 sample3 = GetPaddedSamplesGridSize();
-    
-    [loop]
-    for (uint i = 0; i < 8; i++)
-    {
-        uint3 pos = key.CellCoord + GetCornerOffset(i);
-        float corner = DensityMap[GetDensitySampleIndexLocal(pos, key.ChunkKeyIndex, sample3)];
-        
-        if (corner < ISOLevel)
-            CellCoord |= (1u << i);
-    }
-
-    if (CellCoord == 0 || CellCoord == 255)
-        return;
-    
-    uint localCount = 0;
-    
-    [loop]
-    for (int i = 0; i < 16; i += 3)
-    {
-        int a = GetTriangleEdgeIndex(CellCoord, i + 0);
-        int b = GetTriangleEdgeIndex(CellCoord, i + 1);
-        int c = GetTriangleEdgeIndex(CellCoord, i + 2);
-
-        if (a == -1 || b == -1 || c == -1)
-            break;
-        
-        localCount++;
-    }
-    
-    InterlockedAdd(TriangleCount[key.ChunkKeyIndex], localCount);
-}
-
-void March(ChunkCellContext key)
-{
-    uint CellCoord = 0;
-    float corner[8];
-    float3 cornerPos[8];
-    
-    int step = GetCellStep(key.Chunk.LodIndex);
-    float3 world = ChunkOriginToWorld(key.Chunk.Origin);
-    int3 sample3 = GetPaddedSamplesGridSize();
-    
-    [loop]
-    for (uint i = 0; i < 8; i++)
-    {
-        uint3 pos = key.CellCoord + GetCornerOffset(i);
-        corner[i] = DensityMap[GetDensitySampleIndexLocal(pos, key.ChunkKeyIndex, sample3)];
-        cornerPos[i] = float3(pos) * step + world;
-        
-        if (corner[i] < ISOLevel)
-            CellCoord |= (1u << i);
-    }
-
-    if (CellCoord == 0 || CellCoord == 255)
-        return;
-    
-    float3 cornerNormal[8];
-    
-    [loop]
-    for (int i = 0; i < 8; i++)
-    {
-        int3 pos = key.CellCoord + GetCornerOffset(i);
-        cornerNormal[i] = GetNormalFast(pos, key.ChunkKeyIndex, sample3);
-    }
-    
-    /*
-    ***
-    
-    If you want even flatter
-    float3 normal = normalize(cross(worldB - worldA, worldC - worldA));
-    */
-    
-    /*
-    
-    ChatGPT helped me with this fuckery. In my previous instance I was making calls to the DensityMap
-    itself. Using this below float increased performance by quite a bit, and the lighting looks so much better.
-    
-    What we're doing is taking the opposite face of the cube and sum them up.
-    X - (1-0), (3-2), (5-4), (7-6)
-    Y - (2-0), (3-1), (6-4), (7-5)
-    Z - (4-0), (5-1), (6-2), (7-3)
-    
-        y
-        |
-        4 ---- 5
-        /|     /|
-        0 ---- 1|
-        | 6 --|- 7 ---> x
-        |/    |/
-        2 ----3
-        /
-        z
-
-    Approximate central difference from cube corners 
-    float3 grad = float3(
-        corner[1] - corner[0] + corner[3] - corner[2] + corner[5] - corner[4] + corner[7] - corner[6],
-        corner[2] - corner[0] + corner[3] - corner[1] + corner[6] - corner[4] + corner[7] - corner[5],
-        corner[4] - corner[0] + corner[5] - corner[1] + corner[6] - corner[2] + corner[7] - corner[3]);
-    float3 normal = normalize(grad);
-    */  
-
-    [loop]
-    for (int i = 0; i < 16; i += 3)
-    {
-        int a = GetTriangleEdgeIndex(CellCoord, i + 0);
-        int b = GetTriangleEdgeIndex(CellCoord, i + 1);
-        int c = GetTriangleEdgeIndex(CellCoord, i + 2);
-
-        if (a == -1 || b == -1 || c == -1)
-            break;
-
-        int2 edge0 = GetEdgeConnection(a);
-        int2 edge1 = GetEdgeConnection(b);
-        int2 edge2 = GetEdgeConnection(c);
-        
-        // If you like a more minecraft look make RCP its own dedicated variable.
-        // I was doing that for quite some time, I liked the look but for some
-        // reason bringing them together makes smooth terrain :o
-        float t0 = (corner[edge0.x] - ISOLevel) * rcp(corner[edge0.x] - corner[edge0.y]);
-        float t1 = (corner[edge1.x] - ISOLevel) * rcp(corner[edge1.x] - corner[edge1.y]);
-        float t2 = (corner[edge2.x] - ISOLevel) * rcp(corner[edge2.x] - corner[edge2.y]);
-        
-        float3 worldA = lerp(cornerPos[edge0.x], cornerPos[edge0.y], t0);
-        float3 worldB = lerp(cornerPos[edge1.x], cornerPos[edge1.y], t1);
-        float3 worldC = lerp(cornerPos[edge2.x], cornerPos[edge2.y], t2);
-        
-        float3 normA = normalize(lerp(cornerNormal[edge0.x], cornerNormal[edge0.y], t0));
-        float3 normB = normalize(lerp(cornerNormal[edge1.x], cornerNormal[edge1.y], t1));
-        float3 normC = normalize(lerp(cornerNormal[edge2.x], cornerNormal[edge2.y], t2));
-        
-        TriangleData tri;
-        tri.a = worldA;
-        tri.b = worldB;
-        tri.c = worldC;
-        tri.NormalA = normA;
-        tri.NormalB = normB;
-        tri.NormalC = normC;
-        tri.ChunkKeyIndex = key.ChunkKeyIndex;
-        tri.LodIndex = key.Chunk.LodIndex;
-        
-        uint next;
-        InterlockedAdd(TriangleCursor[key.ChunkKeyIndex], 1, next);
-        uint index = key.Chunk.SourceOffset + next;
-        TriangleSourceBuffer[index] = tri;
-    }
-}
-
-[numthreads(4, 4, 4)]
-void RunMarchingCubesPrePass(uint3 id : SV_DispatchThreadID)
-{
-    ChunkCellContext access = GetChunkCellCubes(id, Offset, ChunkInputs);
-    CountTriangles(access);
-}
-
-// MarchingCubesPass
-// Main meshing kernel. Reads voxel density data for a given chunk, determines
-// which edges of the cube intersect the ISO surface, and emits triangles
-// according to the marching cubes lookup tables.
-// 
-// Each thread processes one cube in the chunk's voxel grid.
-[numthreads(4, 4, 4)]
-void RunMarchingCubes(uint3 id : SV_DispatchThreadID)
-{
-    ChunkCellContext access = GetChunkCellCubes(id, Offset, ChunkInputs);
-    March(access);
-}
