+++ Terrain/Engine/Tables/TransvoxelTables.cs
@@ -5,10 +5,41 @@
  * The goal here is not to re-invent Transvoxel, but to adapt the tables and data
  * into a GPU-friendly format that works cleanly with Unity compute shaders.
  */
+using System.Collections.Generic;
+using System.Runtime.InteropServices;
+using UnityEngine;
+
 namespace GingerVoxelSystem.Engine.Helpers
 {
-    using UnityEngine;
+    /// <summary>
+    /// GPU-friendly version of regular cell data.
+    /// This mirrors the CPU-side Transvoxel tables as closely as possible,
+    /// but stays within HLSL / StructuredBuffer limitations.
+    /// </summary>
+    public struct RegularCellDataGPU
+    {
+        public int VertexCount;
+        public int TriangleCount;
 
+        public int IndicesStart;
+        public int IndicesCount;
+    }
+
+    /// <summary>
+    /// Describes a range into the packed vertex data buffer.
+    /// The original Transvoxel vertex data is jagged, so it must be flattened
+    /// before it can be accessed cleanly from HLSL.
+    /// </summary>
+    public struct VertexData
+    {
+        public uint VertexStart;
+        public uint VertexCount;
+    }
+
+    /// <summary>
+    /// Source
+    /// https://github.com/bbQsauce5/transvoxel-unity/blob/main/Runtime/Mesher/TransvoxelTables.cs
+    /// </summary>
     public class RegularCellData
     {
         byte geometryCounts;
@@ -41,38 +72,12 @@ namespace GingerVoxelSystem.Engine.Helpers
         }
     }
 
-    public class TransvoxelTables
+    /// <summary>
+    /// Source
+    /// https://github.com/bbQsauce5/transvoxel-unity/blob/main/Runtime/Mesher/TransvoxelTables.cs
+    /// </summary>
+    public static class TransvoxelTables
     {
-        // Regular data.
-        public static readonly Vector3Int[] RegularCornerOffset =
-        {
-            new Vector3Int(0,0,0), // 0         6-------7
-		    new Vector3Int(1,0,0), // 1        /|      /|
-		    new Vector3Int(0,0,1), // 2	      / |     / |  Corners
-		    new Vector3Int(1,0,1), // 3	     4-------5  |
-		    new Vector3Int(0,1,0), // 4	     |  2----|--3
-		    new Vector3Int(1,1,0), // 5	     | /     | /   y z
-		    new Vector3Int(0,1,1), // 6	     |/      |/    |/
-		    new Vector3Int(1,1,1)  // 7	     0-------1     o--x
-	    };
-
-        public static readonly Vector3Int[] TransitionCornerOffset =
-        {
-            new Vector3Int(0,0,0), // 0	         B-----------C
-		    new Vector3Int(1,0,0), // 1         /|          /|
-		    new Vector3Int(2,0,0), // 2        / |         / |	         
-		    new Vector3Int(0,1,0), // 3       /  |        /  |                       
-		    new Vector3Int(1,1,0), // 4      6-----7-----8   |	    Corners      	   
-		    new Vector3Int(2,1,0), // 5      |   | |     |   |
-		    new Vector3Int(0,2,0), // 6      |   9 |-----|-- A      y z	
-		    new Vector3Int(1,2,0), // 7      3-----4-----5  /       |/       
-		    new Vector3Int(2,2,0), // 8      | /   |     | /        o--x	        	   
-		    new Vector3Int(0,0,2), // 9      |/    |  	 |/	           
-		    new Vector3Int(2,0,2), // A	     0-----1-----2               
-		    new Vector3Int(0,2,2), // B	   
-		    new Vector3Int(2,2,2)  // C
-	    };
-
         public static readonly int[] RegularCellClass =
         {
             0,  1,  1,  3,  1,  3,  2,  4,  1,  2,  3,  4,  3,  4,  4,  3,
@@ -468,11 +473,6 @@ namespace GingerVoxelSystem.Engine.Helpers
             new RegularCellData(0x97, new byte[] {0, 1, 4, 1, 3, 4, 1, 2, 3, 2, 5, 6, 2, 6, 3, 0, 7, 8, 0, 4, 7}),
             new RegularCellData(0xA8, new byte[] {0, 1, 5, 1, 4, 5, 1, 2, 4, 2, 3, 4, 3, 6, 7, 3, 7, 4, 0, 8, 9, 0, 5, 8}),
             new RegularCellData(0xA8, new byte[] {0, 1, 5, 1, 4, 5, 1, 2, 4, 2, 3, 4, 2, 6, 3, 3, 6, 7, 0, 8, 9, 0, 5, 8})
-        };
-
-        public static readonly int[] TransitionCornerData =
-        {
-            48, 33, 32, 18, 64, 130, 16, 129, 128, 55, 39, 23, 135
         };
 
         public static readonly ushort[][] TransitionVertexData = new ushort[][]
@@ -990,5 +990,291 @@ namespace GingerVoxelSystem.Engine.Helpers
             new ushort[] {0x2301, 0x1503, 0x199B, 0x289A},
             new ushort[] {}
         };
+    }
+
+    /// <summary>
+    /// Owns and manages all GPU lookup tables required by the Transvoxel algorithm.
+    /// This class is responsible for creating, populating, binding, and disposing
+    /// all related ComputeBuffers.
+    /// </summary>
+    public static class TransvoxelGpuTables
+    {
+        /// <summary>
+        /// Tracks whether the regular Transvoxel buffers have been initialized.
+        /// </summary>
+        private static bool isInitialized = false;
+
+        /*
+         * ===============================================================
+         * Regular Transvoxel lookup tables
+         * ===============================================================
+         */
+        private static ComputeBuffer RegularCellClass;
+
+        // RegularCellData
+        private static ComputeBuffer RegularCellTable;
+        private static ComputeBuffer RegularCellIndices;
+
+        // RegularVertexData
+        private static ComputeBuffer RegularVertexRanges;
+        private static ComputeBuffer RegularVertexData;
+
+        /*
+         * ===============================================================
+         * Transition Transvoxel lookup tables (WIP)
+         * ===============================================================
+         */
+        private static ComputeBuffer TransitionCellClass;
+
+        // TransitionRegularCellData
+        private static ComputeBuffer TransitionCellTable;
+        private static ComputeBuffer TransitionCellIndices;
+
+        // TransitionVertexData
+        private static ComputeBuffer TransitionVertexRanges;
+        private static ComputeBuffer TransitionVertexData;
+
+        /// <summary>
+        /// Binds all required Transvoxel buffers to a compute shader kernel.
+        /// Buffers are lazily initialized on first use.
+        /// </summary>
+        /// <param name="shader"></param>
+        /// <param name="kernelId"></param>
+        public static void SetBuffer(ComputeShader shader, int kernelId)
+        {
+            if (!isInitialized)
+            {
+                LoadRegular();
+                LoadTransition();
+                isInitialized = true;
+            }
+
+            shader.SetBuffer(kernelId, "RegularCellClass", RegularCellClass);
+            shader.SetBuffer(kernelId, "RegularCellTable", RegularCellTable);
+            shader.SetBuffer(kernelId, "RegularCellIndices", RegularCellIndices);
+            shader.SetBuffer(kernelId, "RegularVertexRanges", RegularVertexRanges);
+            shader.SetBuffer(kernelId, "RegularVertexData", RegularVertexData);
+
+            shader.SetBuffer(kernelId, "TransitionCellClass", TransitionCellClass);
+            shader.SetBuffer(kernelId, "TransitionCellTable", TransitionCellTable);
+            shader.SetBuffer(kernelId, "TransitionCellIndices", TransitionCellIndices);
+            shader.SetBuffer(kernelId, "TransitionVertexRanges", TransitionVertexRanges);
+            shader.SetBuffer(kernelId, "TransitionVertexData", TransitionVertexData);
+        }
+
+        /// <summary>
+        /// Releases all Transvoxel-related GPU buffers.
+        /// Should be called on shutdown or domain reload.
+        /// </summary>
+        public static void Dispose()
+        {
+            ReleaseBuffer(ref RegularCellClass);
+            ReleaseBuffer(ref RegularCellTable);
+            ReleaseBuffer(ref RegularCellIndices);
+            ReleaseBuffer(ref RegularVertexRanges);
+            ReleaseBuffer(ref RegularVertexData);
+            ReleaseBuffer(ref TransitionCellClass);
+            ReleaseBuffer(ref TransitionCellTable);
+            ReleaseBuffer(ref TransitionCellIndices);
+            ReleaseBuffer(ref TransitionVertexRanges);
+            ReleaseBuffer(ref TransitionVertexData);
+        }
+
+        #region Regular
+        /// <summary>
+        /// Allocates and uploads all regular Transvoxel lookup tables to the GPU.
+        /// </summary>
+        private static void LoadRegular()
+        {
+            int v3Size = Marshal.SizeOf<Vector3Int>();
+
+            RegularCellClass = new ComputeBuffer(256, sizeof(int));
+
+            // Regular Cell Data.
+            RegularCellTable = new ComputeBuffer(TransvoxelTables.RegularCellData.Length, Marshal.SizeOf<RegularCellDataGPU>());
+            RegularCellIndices = new ComputeBuffer(156, Marshal.SizeOf<uint>());
+
+            // Regular Vertex Data
+            RegularVertexRanges = new ComputeBuffer(TransvoxelTables.RegularVertexData.Length, Marshal.SizeOf<VertexData>());
+            RegularVertexData = new ComputeBuffer(1536, sizeof(uint));
+
+            RegularCellClass.SetData(TransvoxelTables.RegularCellClass);
+
+            // RegularCellData
+            LoadRegularCellTable();
+
+            // RegularVertexData
+            LoadRegularVertexData();
+        }
+
+        /// <summary>
+        /// Flattens and uploads regular cell topology data into GPU buffers.
+        /// </summary>
+        private static void LoadRegularCellTable()
+        {
+            List<RegularCellDataGPU> dataGPU = new();
+            List<uint> indices = new();
+
+            int startIndex = 0;
+
+            foreach (var data in TransvoxelTables.RegularCellData)
+            {
+                RegularCellDataGPU newData = new RegularCellDataGPU();
+                newData.VertexCount = (int)data.GetVertexCount();
+                newData.TriangleCount = (int)data.GetTriangleCount();
+                newData.IndicesStart = startIndex;
+                newData.IndicesCount = newData.TriangleCount * 3;
+
+                var lindices = data.GetIndices();
+                foreach (var indice in lindices)
+                {
+                    indices.Add(indice);
+                }
+
+                startIndex += newData.IndicesCount;
+                dataGPU.Add(newData);
+            }
+
+            Debug.Assert(indices.Count == 156);
+
+            RegularCellTable.SetData(dataGPU.ToArray());
+            RegularCellIndices.SetData(indices.ToArray());
+        }
+
+        /// <summary>
+        /// Flattens and uploads packed vertex descriptor data into GPU buffers.
+        /// </summary>
+        private static void LoadRegularVertexData()
+        {
+            List<VertexData> vertexRanges = new();
+            List<uint> vertexData = new();
+
+            uint start = 0;
+            foreach (var vertexList in TransvoxelTables.RegularVertexData)
+            {
+                VertexData newData = new VertexData();
+                newData.VertexStart = start;
+                newData.VertexCount = (uint)vertexList.Length;
+
+                vertexRanges.Add(newData);
+
+                // Add packed vertex descriptors
+                foreach (ushort v in vertexList)
+                {
+                    vertexData.Add((uint)v);
+                }
+
+                start += newData.VertexCount;
+            }
+
+            Debug.Assert(vertexData.Count == 1536);
+            Debug.Assert(vertexRanges.Count == TransvoxelTables.RegularVertexData.Length);
+
+            RegularVertexRanges.SetData(vertexRanges.ToArray());
+            RegularVertexData.SetData(vertexData.ToArray());
+        }
+        #endregion
+
+        #region Transition
+        private static void LoadTransition()
+        {
+            // MAY need to be converted to int.
+            TransitionCellClass = new ComputeBuffer(TransvoxelTables.TransitionCellClass.Length, sizeof(int));
+
+            // Transition Cell Data.
+            TransitionCellTable = new ComputeBuffer(TransvoxelTables.TransitionRegularCellData.Length, Marshal.SizeOf<RegularCellDataGPU>());
+            TransitionCellIndices = new ComputeBuffer(924, Marshal.SizeOf<uint>());
+
+            // Transition Vertex Data
+            TransitionVertexRanges = new ComputeBuffer(TransvoxelTables.TransitionVertexData.Length, Marshal.SizeOf<VertexData>());
+            TransitionVertexData = new ComputeBuffer(4096, sizeof(uint));
+
+            TransitionCellClass.SetData(TransvoxelTables.TransitionCellClass);
+
+            LoadTransitionCellTable();
+            LoadTransitionVertexData();
+        }
+
+        /// <summary>
+        /// Flattens and uploads regular cell topology data into GPU buffers.
+        /// </summary>
+        private static void LoadTransitionCellTable()
+        {
+            List<RegularCellDataGPU> dataGPU = new();
+            List<uint> indices = new();
+
+            int startIndex = 0;
+
+            foreach (var data in TransvoxelTables.TransitionRegularCellData)
+            {
+                RegularCellDataGPU newData = new RegularCellDataGPU();
+                newData.VertexCount = (int)data.GetVertexCount();
+                newData.TriangleCount = (int)data.GetTriangleCount();
+                newData.IndicesStart = startIndex;
+                newData.IndicesCount = newData.TriangleCount * 3;
+
+                var lindices = data.GetIndices();
+                foreach (var indice in lindices)
+                {
+                    indices.Add(indice);
+                }
+
+                startIndex += newData.IndicesCount;
+                dataGPU.Add(newData);
+            }
+
+            Debug.Assert(indices.Count == 924);
+
+            TransitionCellTable.SetData(dataGPU.ToArray());
+            TransitionCellIndices.SetData(indices.ToArray());
+        }
+
+        /// <summary>
+        /// Flattens and uploads packed vertex descriptor data into GPU buffers.
+        /// </summary>
+        private static void LoadTransitionVertexData()
+        {
+            List<VertexData> vertexRanges = new();
+            List<uint> vertexData = new();
+
+            uint start = 0;
+            foreach (var vertexList in TransvoxelTables.TransitionVertexData)
+            {
+                VertexData newData = new VertexData();
+                newData.VertexStart = start;
+                newData.VertexCount = (uint)vertexList.Length;
+
+                vertexRanges.Add(newData);
+
+                // Add packed vertex descriptors
+                foreach (ushort v in vertexList)
+                {
+                    vertexData.Add((uint)v);
+                }
+
+                start += newData.VertexCount;
+            }
+
+            Debug.Assert(vertexData.Count == 4096);
+            Debug.Assert(vertexRanges.Count == TransvoxelTables.TransitionVertexData.Length);
+
+            TransitionVertexRanges.SetData(vertexRanges.ToArray());
+            TransitionVertexData.SetData(vertexData.ToArray());
+        }
+        #endregion
+
+        /// <summary>
+        /// Safely releases and disposes a ComputeBuffer.
+        /// </summary>
+        /// <param name="buffer"></param>
+        private static void ReleaseBuffer(ref ComputeBuffer buffer)
+        {
+            if (buffer != null)
+            {
+                buffer.Release();
+                buffer.Dispose();
+                buffer = null;
+            }
+        }
     }
 }
